Return Upcaster to original location
