+++ examples/common/b.rs
@@ -38,4 +38,4 @@ pub struct EventB {
 }
 
 #[cfg(feature = "upcasting")]
-impl esrs::sql::event::Upcaster for EventB {}
+impl esrs::event::Upcaster for EventB {}
