+++ src/sql/event.rs
@@ -10,23 +10,6 @@ use crate::types::SequenceNumber;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
 
-#[cfg(feature = "upcasting")]
-pub trait Upcaster
-where
-    Self: Sized,
-{
-    fn upcast(value: serde_json::Value, _version: Option<i32>) -> Result<Self, serde_json::Error>
-    where
-        Self: DeserializeOwned,
-    {
-        serde_json::from_value(value)
-    }
-
-    fn current_version() -> Option<i32> {
-        None
-    }
-}
-
 #[cfg(not(feature = "upcasting"))]
 pub trait Persistable: Serialize + DeserializeOwned {}
 
@@ -34,10 +17,10 @@ pub trait Persistable: Serialize + DeserializeOwned {}
 impl<T> Persistable for T where T: Serialize + DeserializeOwned {}
 
 #[cfg(feature = "upcasting")]
-pub trait Persistable: Serialize + DeserializeOwned + Upcaster {}
+pub trait Persistable: Serialize + DeserializeOwned + crate::event::Upcaster {}
 
 #[cfg(feature = "upcasting")]
-impl<T> Persistable for T where T: Serialize + DeserializeOwned + Upcaster {}
+impl<T> Persistable for T where T: Serialize + DeserializeOwned + crate::event::Upcaster {}
 
 /// Event representation on the event store
 #[derive(sqlx::FromRow, serde::Serialize, serde::Deserialize, Debug)]
